Fix rename parser to avoid name collision
